@@ -1,12 +1,12 @@
 package com.foodie.backend.dto;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
-;
+import com.foodie.backend.model.Role;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 
 public class UserDTO {
+
     private String id;
 
     @NotBlank(message = "Name is required")
@@ -17,11 +17,26 @@ public class UserDTO {
     @Email(message = "Please provide a valid email address")
     private String email;
 
+    private String address;
+    private String phone;
+    private Role role;  // ✅ Include role (enum)
 
+    // --- Getters and Setters ---
     public String getId() { return id; }
     public void setId(String id) { this.id = id; }
+
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
+
     public String getEmail() { return email; }
     public void setEmail(String email) { this.email = email; }
+
+    public String getAddress() { return address; }
+    public void setAddress(String address) { this.address = address; }
+
+    public String getPhone() { return phone; }
+    public void setPhone(String phone) { this.phone = phone; }
+
+    public Role getRole() { return role; }
+    public void setRole(Role role) { this.role = role; }
 }
