@@ -1,7 +1,8 @@
 package com.foodie.backend.dto;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 
 public class UserLoginDTO {
     @NotBlank(message = "Email is required")
