@@ -1,61 +1,83 @@
 package com.foodie.backend.model;
 
-
 import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.List;
-
 
 @Document(collection = "restaurants")
 public class Restaurant {
+
     @Id
     private String id;
     private String name;
     private String address;
     private String cuisineType;
+    private String ownerId;
 
-    public String getId() {
-        return id;
+    private Boolean approved = false;   // ✅ Default: not yet approved
+    private Boolean suspended = false;  // ✅ Admin can suspend
+
+    private List<MenuItem> menuItems = new ArrayList<>();
+
+    // ----------------- Restaurant Fields -----------------
+    public String getId() { return id; }
+    public void setId(String id) { this.id = id; }
+
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+
+    public String getAddress() { return address; }
+    public void setAddress(String address) { this.address = address; }
+
+    public String getCuisineType() { return cuisineType; }
+    public void setCuisineType(String cuisineType) { this.cuisineType = cuisineType; }
+
+    public String getOwnerId() { return ownerId; }
+    public void setOwnerId(String ownerId) { this.ownerId = ownerId; }
+
+    public List<MenuItem> getMenuItems() { return menuItems; }
+    public void setMenuItems(List<MenuItem> menuItems) { this.menuItems = menuItems; }
+
+    public Boolean getApproved() {
+        return approved;
     }
 
-    public void setId(String id) {
-        this.id = id;
+    public void setApproved(Boolean approved) {
+        this.approved = approved;
     }
 
-    public String getName() {
-        return name;
+    public Boolean getSuspended() {
+        return suspended;
     }
 
-    public void setName(String name) {
-        this.name = name;
+    public void setSuspended(Boolean suspended) {
+        this.suspended = suspended;
     }
 
-    public String getAddress() {
-        return address;
+    // ----------------- Inner MenuItem -----------------
+    public static class MenuItem {
+        private String id;
+        private String name;
+        private String description;
+        private BigDecimal price;
+        private String category;
+
+        public String getId() { return id; }
+        public void setId(String id) { this.id = id; }
+
+        public String getName() { return name; }
+        public void setName(String name) { this.name = name; }
+
+        public String getDescription() { return description; }
+        public void setDescription(String description) { this.description = description; }
+
+        public BigDecimal getPrice() { return price; }
+        public void setPrice(BigDecimal price) { this.price = price; }
+
+        public String getCategory() { return category; }
+        public void setCategory(String category) { this.category = category; }
     }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getCuisineType() {
-        return cuisineType;
-    }
-
-    public void setCuisineType(String cuisineType) {
-        this.cuisineType = cuisineType;
-    }
-
-    public List<String> getMenuItems() {
-        return menuItems;
-    }
-
-    public void setMenuItems(List<String> menuItems) {
-        this.menuItems = menuItems;
-    }
-
-    private List<String> menuItems;
-
 }
