@@ -1,15 +1,12 @@
 package com.foodie.backend.service;
 
-import com.foodie.backend.dto.UserDTO;
-import com.foodie.backend.dto.UserLoginDTO;
-import com.foodie.backend.dto.UserRegistrationDTO;
+import com.foodie.backend.dto.*;
 
 import java.util.List;
 
 public interface UserService {
-    UserDTO registerUser(UserRegistrationDTO registrationDto);
-    UserDTO loginUser(UserLoginDTO loginDto);
-
+    AuthResponseDTO registerUser(UserRegistrationDTO registrationDto);
+    AuthResponseDTO loginUser(UserLoginDTO loginDto);
     UserDTO getUserById(String id);
     List<UserDTO> getAllUsers();
     UserDTO updateUser(String id, UserDTO userDto);
