@@ -1,92 +1,135 @@
 package com.foodie.backend.service;
 
-import com.foodie.backend.dto.UserDTO;
-import com.foodie.backend.dto.UserLoginDTO;
-import com.foodie.backend.dto.UserRegistrationDTO;
+import com.foodie.backend.dto.*;
 import com.foodie.backend.mapper.UserMapper;
+import com.foodie.backend.model.Role;
 import com.foodie.backend.model.User;
 import com.foodie.backend.repository.UserRepository;
+import com.foodie.backend.utils.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Instant;
 import java.util.List;
 
 @Service
 @Transactional
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
+    private final PasswordEncoder passwordEncoder;
+    private final AuthenticationManager authenticationManager;
+    private final JwtUtil jwtUtil;
+
     @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private UserMapper userMapper;
-    @Autowired
-    private PasswordEncoder passwordEncoder;
+    public UserServiceImpl(UserRepository userRepository,
+                           UserMapper userMapper,
+                           PasswordEncoder passwordEncoder,
+                           AuthenticationManager authenticationManager,
+                           JwtUtil jwtUtil) {
+        this.userRepository = userRepository;
+        this.userMapper = userMapper;
+        this.passwordEncoder = passwordEncoder;
+        this.authenticationManager = authenticationManager;
+        this.jwtUtil = jwtUtil;
+    }
+
     @Override
-    public UserDTO registerUser(UserRegistrationDTO userRegistrationDTO){
-        if(userRepository.existsByEmail(userRegistrationDTO.getEmail())){
+    public AuthResponseDTO registerUser(UserRegistrationDTO registrationDto) {
+        if (userRepository.existsByEmail(registrationDto.getEmail())) {
             throw new RuntimeException("Email already in use");
         }
-        User user = userMapper.toEntity(userRegistrationDTO);
-        user.setPassword(passwordEncoder.encode(userRegistrationDTO.getPassword()));
-        user = userRepository.save(user);
-        return userMapper.toDto(user);
-    }
-    @Override
-    public UserDTO loginUser(UserLoginDTO userLoginDTO){
-        User user = userRepository.findByEmail(userLoginDTO.getEmail())
-                .orElseThrow(() -> new RuntimeException("Invalid Credentials"));
-        if(!passwordEncoder.matches(userLoginDTO.getPassword(), user.getPassword())){
-            throw new RuntimeException("Invalid Credentials");
 
-        }
-        return userMapper.toDto(user);
+        User user = userMapper.toEntity(registrationDto);
+        user.setPassword(passwordEncoder.encode(registrationDto.getPassword()));
+
+        // ✅ Set role using enum properly
+        user.setRole(Role.USER);
+
+        User savedUser = userRepository.save(user);
+        System.out.println("Saved user: " + savedUser); // Debug print
+
+        String token = jwtUtil.generateToken(savedUser.getEmail());
+        System.out.println("Generated token: " + token); // Debug print
+
+        Instant expiration = jwtUtil.extractExpiration(token).toInstant();
+        UserDTO userDto = userMapper.toDto(savedUser);
+
+        return new AuthResponseDTO(token, "Bearer", expiration, userDto);
     }
+
+    @Override
+    public AuthResponseDTO loginUser(UserLoginDTO loginDto) {
+        try {
+            Authentication authentication = authenticationManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(
+                            loginDto.getEmail(),
+                            loginDto.getPassword()
+                    )
+            );
+
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+
+            User user = userRepository.findByEmail(loginDto.getEmail())
+                    .orElseThrow(() -> new RuntimeException("User not found"));
+
+            String token = jwtUtil.generateToken(user.getEmail());
+            Instant expiration = jwtUtil.extractExpiration(token).toInstant();
+            UserDTO userDto = userMapper.toDto(user);
+
+            return new AuthResponseDTO(token, "Bearer", expiration, userDto);
+        } catch (Exception e) {
+            throw new RuntimeException("Login failed: " + e.getMessage());
+        }
+    }
+
     @Override
     public UserDTO getUserById(String id) {
-        User user = userRepository.findById(id).
-                orElseThrow(()->new RuntimeException("User Not Found"));
-
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
         return userMapper.toDto(user);
     }
 
     @Override
     public List<UserDTO> getAllUsers() {
-        List<User> all = userRepository.findAll();
-        return all.stream().
-                map(userMapper::toDto)
+        return userRepository.findAll().stream()
+                .map(userMapper::toDto)
                 .toList();
-
     }
 
     @Override
     public UserDTO updateUser(String id, UserDTO userDto) {
-
-        User found = userRepository.findById(id)
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
-        found.setName(userDto.getName());
-        found.setEmail(userDto.getEmail());
-        User save = userRepository.save(found);
-        return userMapper.toDto(save);
+        user.setName(userDto.getName());
+        user.setEmail(userDto.getEmail());
+        user.setAddress(userDto.getAddress());
+        user.setPhone(userDto.getPhone());
+
+        User updatedUser = userRepository.save(user);
+        return userMapper.toDto(updatedUser);
     }
 
     @Override
     public void deleteUser(String id) {
-
-         userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found");
+        }
         userRepository.deleteById(id);
-
-
-
     }
 
     @Override
     public UserDTO findByEmail(String email) {
         User user = userRepository.findByEmail(email)
-                .orElseThrow(()->new RuntimeException("User Not Found"));
+                .orElseThrow(() -> new RuntimeException("User not found"));
         return userMapper.toDto(user);
     }
 
